+++ core/src/filed/fileset.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2000-2011 Free Software Foundation Europe e.V.
    Copyright (C) 2011-2012 Planets Communications B.V.
-   Copyright (C) 2013-2020 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2022 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -110,9 +110,11 @@ static void append_file(JobControlRecord* jcr,
  * Add fname to include/exclude fileset list. First check for
  * | and < and if necessary perform command.
  */
-void AddFileToFileset(JobControlRecord* jcr, const char* fname, bool IsFile)
+void AddFileToFileset(JobControlRecord* jcr,
+                      const char* fname,
+                      bool IsFile,
+                      findFILESET* fileset)
 {
-  findFILESET* fileset = jcr->impl->ff->fileset;
   char* p;
   Bpipe* bpipe;
   POOLMEM* fn;
@@ -298,11 +300,11 @@ void AddFileset(JobControlRecord* jcr, const char* item)
       break;
     case 'F': /* File */
       state = state_include;
-      AddFileToFileset(jcr, item, true);
+      AddFileToFileset(jcr, item, true, jcr->impl->ff->fileset);
       break;
     case 'P': /* Plugin */
       state = state_include;
-      AddFileToFileset(jcr, item, false);
+      AddFileToFileset(jcr, item, false, jcr->impl->ff->fileset);
       break;
     case 'R': /* Regex */
       state = AddRegexToFileset(jcr, item, subcode);