+++ core/src/lib/parse_conf.h
@@ -353,7 +353,10 @@ class ConfigurationParser {
   BStringList warnings_;
 
 
-  const char* GetDefaultConfigDir();
+ public:
+  static const char* GetDefaultConfigDir();
+
+ private:
   bool GetConfigFile(PoolMem& full_path,
                      const char* config_dir,
                      const char* config_filename);